Add strerror(errno) to error message when fork() fails
Trying to figure out why OS X sometimes fails to fork.  Sadly, I have
a feeling it is due to resource limits imposed by the OS, and even
though child processes exit() we may have to figure out a way reuse
the process for as long as we can, only calling fork() again if the
child dies.