Remove a DLL I never meant to commit
Also, edit .gitignore so it doesn't happen again.

The problem is that mingw uses some DLLs which are not in Wine's path,
so when trying to test the Windows version on Linux it's convenient to
copy the DLL to the local directory.  Not sure how I managed to `git
add` it, but it happened.