longjmp on error instead of aborting in non-forking mode.
This lets us continue with other tests, even when an assertion fails.
It's not perfect; for one thing, it will only work if the error is
triggered in the thread which executes the test.  However, it should
be pretty safe, and usually since it will generally fall back on
abort().