Add munit_errorf_ex
This function is basically the same as
munit_logf_ex(MUNIT_LOG_ERROR, ...), but since it is only called with
errors (which are always fatal) it can be annotated as _Noreturn.  This
is important, since it lets static analyzers know that failing an
assertion is fatal, thus it is safe to assume that code which follows
the assertion must have passed it.  For example, if we assert that
something is not equal to NULL, the compiler/analyzer knows that we can
safely pass the value to a function with arguments annotated as
non-null.