Silence MSVC C4127 in munit.h instead of making the user do it
MSVC C4127 is "conditional expression is constant", which MSVC sounds
like a reasonable error, but it includes code like:

  do {
    ...
  } while (0); // <- here

Which is an extremely common pattern in preprocessor macros which
allows them to de used in many situations where they otherwise couldn't
be, including in µnit's assertion macros.

This patch gets around the issue by doing something like this:

  do {
    ...
    __pragma(warning(push))
    __pragma(warning(disable:4127))
  } while (0)
  __pragma(warning(pop))

Of course it's more complicated than that because we have to do it only
for MSVC; __pragma is MSVC-specific (the standard version is _Pragma),
and even if we could use _Pragma the warning stuff is compiler-specific
too.