Don't use <stdbool.h>.
<stdbool.h> breaks AltiVec/VMX support on clang if it is included
before <altivec.h>.  The problem is that <stdbool.h> defines a bool
macro to _Bool, but AltiVec uses types like "vector bool char", which
will be converted to "vector _Bool char", resulting in a build failure.

This is technically a breaking change, but I don't expect it to cause
significant pain unless someone is relying on munit.h to includie
<stdbool.h> for non-µnit code.  The only place it is used in the API
is a return value for the parse_argument callback, which doesn't seem
to be used very often.

In the event that this does break your code, please just include
<stdbool.h> before including munit.h.