Rename MUNIT__*_DISABLE_MSVC_C4127 to MUNIT_POP_DISABLE_MSVC_C4127_
The double-underscores in the old version trigger trigger a
-Wreserved-id-macro diagnostic in clang in C++ mode.  The diagnostic is
off by default and not even part of -Wextra, but it is in a public
header which I'd like to be able to compile cleanly with -Weverything.

C reserves everything starting with a double underscore (or a single
underscore and an uppercase letter).  I'm not sure if C++ really does
reserve everything with a double underscore anywhere, but it doesn't
really matter; the fact that clang generates a warning is reason
enough for the change.  This is a private symbol (hence the double
underscore in the first place), so there is no down-side to moving it.