Replace exit by _exit in child.
See
https://stackoverflow.com/questions/5422831/what-is-the-difference-between-using-exit-exit-in-a-conventional-linux-fo
and section 25.4 in `The Linux Programming Interface` by Michael
Kerrisk.

The `exit` library call was causing our tests to become extremely slow
when running with ASAN on arm64 and ppc64le architectures. Replacing with
`_exit` fixed the issue.
Haven't exactly pinpointed the issue, but the problem is related to the
child and the parent both flushing the parent's stdio buffers and
calling the exit handlers.

Signed-off-by: Mathieu Borderé <mathieu.bordere@canonical.com>